+++ example/lib/widgets/material_equivalents.dart
@@ -93,9 +93,11 @@ class _UIEquivalentsState extends State<UIEquivalents> {
       ],
       [
         const Text('RadioButton'),
-        RadioButton(
-          checked: radioChecked,
-          onChanged: (final v) => setState(() => radioChecked = v),
+        RadioGroup<bool>(
+          onChanged: (final v) =>
+              setState(() => radioChecked = v ?? radioChecked),
+          groupValue: true,
+          child: RadioButton(value: radioChecked),
         ),
         RadioGroup<bool>(
           onChanged: (final v) => setState(() => radioChecked = !radioChecked),